requirements.txt
flask
selenium
webdriver-manager
requests
python-dotenv
